+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and the project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.
 - Add option to view months available as string (e.g., March - November), in addition to the graph.
 - Add option to view times available as string (e.g., 8AM - 7PM), in addition to the graph.
 - Add option to view times as 12-hour (3PM) or 24-hour (15:00) time.
+- Marker on times available graph showing current hour (similar to in-game).
 ### Fixed
 - Fix for off-by-one issue when getting availability, which led to incorrectly showing the months that an item was available.
 - Progress bars on Progress screen will now show even if they are unable to load items.